Pushing Changes to prototypes and Main Page
Combined all of the prototypes onto to the project main app page
USCIS,html. Also fixed all of the issues with the map, such as color
scheme, hover text, adding a legend, included all of the schools in the
state totals. In addition, I added a search function that allows the
user to search and add schools to be compared in the pie chart.